clean up old references and move recordersetting functions back into recordersetting to keep this overall change scoped
