+++ src/AccessibilityInsights.SharedUx/Settings/ConfigurationManager.cs
@@ -115,104 +115,15 @@ namespace AccessibilityInsights.SharedUx.Settings
             var configpath = Path.Combine(DirectoryManagement.sUserDataFolderPath, EventConfigFileName);
             try
             {
-                var rcfg = LoadConfiguration(configpath);
+                var rcfg = RecorderSetting.LoadConfiguration(configpath);
                 this.EventConfig = rcfg;
 
             }
             catch (Exception)
             {
                 RecorderSetting.RemoveConfiguration(configpath);
-                this.EventConfig = LoadConfiguration(configpath);
+                this.EventConfig = RecorderSetting.LoadConfiguration(configpath);
             }
-        }
-
-        public static RecorderSetting LoadConfiguration(string path)
-        {
-            // Get Recorder configuration from local location. but if it is not available, get it from default location. 
-            RecorderSetting config = new RecorderSetting();
-            config = RecorderSetting.LoadFromJSON<RecorderSetting>(path);
-            if (config == null)
-            {
-                config = GetDefaultRecordingConfig();
-                config.SerializeInJSON(path);
-            }
-            else
-            {
-                // check whether there is any new events to be added into configuration. 
-                var events = Axe.Windows.Desktop.Types.EventType.GetInstance();
-                var ms = from e in events.GetKeyValuePairList()
-                         where IsNotInList(e.Key, config.Events)
-                         select e;
-
-                if (ms.Count() != 0)
-                {
-                    foreach (var m in ms)
-                    {
-                        config.Events.Add(new RecordEntitySetting()
-                        {
-                            Id = m.Key,
-                            Name = m.Value,
-                            IsRecorded = false,
-                            Type = RecordEntityType.Event,
-                        });
-                    }
-                    config.SerializeInJSON(path);
-                }
-                config.IsListeningAllEvents = false;
-            }
-
-            return config;
-        }
-
-        /// <summary>
-        /// check whether key exist in the given list. 
-        /// </summary>
-        /// <param name="key"></param>
-        /// <param name="events"></param>
-        /// <returns></returns>
-        private static bool IsNotInList(int key, List<RecordEntitySetting> events)
-        {
-            return (from e in events
-                    where e.Id == key
-                    select e).Count() == 0;
-        }
-
-        /// <summary>
-        /// Get the default Event Recording Configuration
-        /// </summary>
-        /// <returns></returns>
-        public static RecorderSetting GetDefaultRecordingConfig()
-        {
-            RecorderSetting config = new RecorderSetting
-            {
-                // Set Individual Event
-                Events = (from e in Axe.Windows.Desktop.Types.EventType.GetInstance().GetKeyValuePairList()
-                          select new RecordEntitySetting()
-                          {
-                              Type = RecordEntityType.Event,
-                              Id = e.Key,
-                              Name = e.Value,
-                              IsRecorded = false
-                          }).ToList(),
-
-                // Set properties
-                Properties = (from e in Axe.Windows.Core.Types.PropertyType.GetInstance().GetKeyValuePairList()
-                              select new RecordEntitySetting()
-                              {
-                                  Type = RecordEntityType.Property,
-                                  Id = e.Key,
-                                  Name = e.Value,
-                                  IsRecorded = false
-                              }).ToList(),
-
-                // Set Global event
-                IsListeningFocusChangedEvent = true,
-
-                // Individual Event Scope
-                ListenScope = UIAutomationClient.TreeScope.TreeScope_Subtree
-            };
-
-            return config;
         }
 
         /// <summary>
